Feature/model docker (#5)
* initial commit data-science-production

* initial commit data-science-production

* created package

* added setup python file

* added new package names

* fixing quotes

* include package data

* initial package

* optimize imports

* optimize imports

* removed imports from init

* added imports in submodule init

* added requires in setup

* added democontainer

* fixed api host

* json content to featureextractor

* fixed flask url path

* Ignore pyc files

* Upgrade pip when building the docker image

Not strictly required but silences the warning of using an old pip
version.

* Move extractors and models into separate packages

* Add some documentation on running the container

* Reorganize Docker and conda

The docker file was not really efficient to iterate quickly.
Changing the order of things helps.

Also fixed the way we create the conda environment. Switched to a
python 2.7 environment as this is required for the verkiezingen use
case.

Finally, we need to activate the environment and start the app in
the same bash command.

* Add code and documentation for the baseline model

This adds code for the model as created for the stemradar. Can't get
it to load though.

* make working for verkiezingmodel

* cleanup script

* cleanup script

* update demo lm model

* added exception printing